refactor : 컴포넌트 기능에 따라 분리
